### 2020-10-09 #2733 @whedon
[REVIEW]: JASP for Audit: Bayesian Tools for the Auditing Practice
**Submitting author:** @koenderks (<a href="http://orcid.org/0000-0002-5533-9349">Koen Derks</a>)
**Repository:** <a href="https://github.com/koenderks/jaspAudit" target ="_blank">https://github.com/koenderks/jaspAudit</a>
**Version:** v0.16
**Reviewer:** @trstewart212 , @fgeertman

- [x] #3483
- [x] #3461
