### 2020-10-09 #2733 @koenderks
@drvinceknight Can this be the result of the fact that whedon selects the default branch from the repository? I believe this has also happened in the pre-review stage. The branch that the submission is located at is the `JfA-JOSS` branch located at `https://github.com/koenderks/jasp-desktop/tree/JfA-JOSS`.

If this is the case, I believe that

will solve the issue.
