### 2020-10-09 #2733 @drvinceknight
> If this is the case, I believe that
>
> @whedon generate pdf from branch JfA-JOSS
>
> will solve the issue.

Thanks @koenderks! That's done the trick. Fyi you can run those whedon commands as well :)
