### 2020-11-10 #2733 @kthyng
@trstewart212 Understood. @drvinceknight and others (including me) are happy to make it as smooth as possible for you. I recommend reading through the [JOSS reviewer guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). Then, you can use the checklist in this review issue to guide your review process. As you identify issues to bring up, we ask you to open issues in the [software repo](https://github.com/koenderks/jasp-desktop) itself and connect them with this review issue by writing something at the top of each issue like "This issue is related to the [review issue](https://github.com/openjournals/joss-reviews/issues/2733)".

@koenderks When I checked your repo for the above comment, I noticed that issues aren't available for your software. You'll need to go into the settings for your software repo and check a box to turn issues on.
