### 2020-12-11 #2733 @trstewart212
Hi Vince,
I’m afraid it has been just one thing after another and I have not been able to devote any time to this. Hopefully things will ease up a bit.
Trevor

Trevor Stewart
175 West 13th Street, Apt 19D
New York, NY 10011
+1(212)929-0321
+1(917)402-9800 iPhone
trsny@verizon.net
