### 2020-12-16 #2733 @trstewart212
Hi Vince,

Perhaps I just don’t know where to look, but I am unable to find any detailed description on GitHub of what this software does. All I have is a two-page JASP for Audit summary.

Somewhere there ought to be documentation that explains in some detail what the software does from an auditor's (user’s) perspective. Descriptions need to be in English accompanied by technical documentation including the underlying mathematics and some worked examples including software input and output. In the absence of such detailed documentation I  don’t see how it is feasible to form an opinion on the submission.

So I am kind of stuck.

Best,

Trevor

Trevor Stewart
175 West 13th Street, Apt 19D
New York, NY 10011
+1(212)929-0321
+1(917)402-9800 iPhone
trsny@verizon.net
