### 2020-12-18 #2733 @koenderks
Dear Vince, Trevor, and Ferry,

I hope you are all doing well in these last days of 2020. Yesterday a new version of JASP has come out (0.14.1), and I have rebased the code in the submission repository ([branch JfA-JOSS](https://github.com/koenderks/jasp-desktop/tree/JfA-JOSS)) to be the same as the code for the new release. JASP 0.14.1 can be downloaded from the JASP website (www.jasp-stats.org).

![image](https://user-images.githubusercontent.com/25059399/102596791-fc6e3180-4119-11eb-8602-1723f9767b1d.png)

Using this new version I can maybe shed some light on where important documentation can be found in the Audit module in JASP, and how to get started using the Audit module. When you first download JASP and open it, you will see the following welcome screen with, at the top, all the modules that are by default loaded in JASP. Our submission concerns the Audit module, which is not visible by default and has to be selected from the modules menu. The module menu is accessible via the "+" icon in the top right corner.

![image](https://user-images.githubusercontent.com/25059399/102597198-959d4800-411a-11eb-89f0-5447b8124fcc.png)

There, we can select "Audit" to load the Audit module.

![image](https://user-images.githubusercontent.com/25059399/102597259-ab127200-411a-11eb-877c-2fb1fc7c8bf1.png)

The Audit module now becomes available in the top ribbon:

![image](https://user-images.githubusercontent.com/25059399/102597311-c2e9f600-411a-11eb-9b0a-1168ef21274e.png)

Clicking this module icon will show the available analyses in the module. Note that for the planning and evaluation analyses, no data file is required by default. The other analyses only become available when a data set is loaded in.

![image](https://user-images.githubusercontent.com/25059399/102597430-ee6ce080-411a-11eb-8cb6-be3874672954.png)

Each analysis has a help file that explains the purpose of the analysis and all available options. It can be found by clicking the "i"' icon on top of the specific analysis form. Several options also have help files that are specifically tailored to those options. These can be found by clicking on the "i" icon next to that specific option.

![image](https://user-images.githubusercontent.com/25059399/102597751-5b807600-411b-11eb-95d7-2aa37e70b99a.png)

Worked out examples of analyses are also included in JASP. These can be found by clicking the menu button in the top left, hovering on "Open", then hovering on "Data Library", and finally clicking "7. Audit".

![image](https://user-images.githubusercontent.com/25059399/102597933-997d9a00-411b-11eb-9886-b89af0cd0d2c.png)

This will bring up the data library for the Audit module. Each entry in this data library has two files associated with it. The left file is an example file, and the right file is the corresponding raw data file. Clicking the (left) example file will load an existing analysis including results based on the corresponding data file. Clicking the (right) data file will simply open the data set.

![image](https://user-images.githubusercontent.com/25059399/102598415-46f0ad80-411c-11eb-984d-65ed51022b05.png)

I hope this provides some guidance on how to get started with JASP for Audit. If you have any questions, do not hesitate to contact me!

Happy holidays,
Koen

P.s. The files for these analyses can be found in the submission repository ([branch JfA-JOSS](https://github.com/koenderks/jasp-desktop/tree/JfA-JOSS)) under:

QML files for the interface:

Resources -> Audit -> qml -> auditBayesianWorkflow.qml
Resources -> Audit -> qml -> auditClassicalWorkflow.qml
Resources -> Audit -> qml -> auditBayesianPlanning.qml
Resources -> Audit -> qml -> auditClassicalPlanning.qml
Resources -> Audit -> qml -> auditSelection.qml
Resources -> Audit -> qml -> auditBayesianEvaluation.qml
Resources -> Audit -> qml -> auditClassicalEvaluation.qml

R files:

JASP-Engine -> JASP -> R -> auditBayesianWorkflow.R
JASP-Engine -> JASP -> R -> auditClassicalWorkflow.R
JASP-Engine -> JASP -> R -> auditBayesianPlanning.R
JASP-Engine -> JASP -> R -> auditClassicalPlanning.R
JASP-Engine -> JASP -> R -> auditSelection.R
JASP-Engine -> JASP -> R -> auditBayesianEvaluation.R
JASP-Engine -> JASP -> R -> auditClassicalEvaluation.R
JASP-Engine -> JASP -> R -> auditCommonFunctions.R
JASP-Engine -> JASP -> R -> auditCommonFunctionBayesian.R
