### 2020-12-18 #2733 @trstewart212
Hi Koen,
This is helpful and I made a bit of progress, but I need help. Just to try it out, I ran the Evaluation option against the overstatement test dataset of 3,500 items. I made up performance materiality of 1%, population size of 100,000 and value of $10,000,000. I am now looking at the screen below. Some questions:
How do we get a negative total tainting/most likely error and a negative 95% upper bound in a sample for overstatement?
Where can I see a detailed explanation of this calculation including the mathematics?
The note says that the upper bound is calculated according to the Stringer method with LTA adjustment. As a reviewer, how do I know this has been correctly implemented, mathematically?
The evaluation refers to prior information. What is the prior, and where does it come from?
Thanks,

Trevor

Trevor Stewart
