### 2020-12-18 #2733 @koenderks
This is my screen with the abovementioned options, yours is not the same I imagine?

![image](https://user-images.githubusercontent.com/25059399/102631703-5983da80-414e-11eb-8379-b847a80db5cd.png)
