### 2020-12-18 #2733 @trstewart212
My screen is the same. I don’t know why the paste didn’t work.

Sent from my iPhone

Trevor Stewart

> On Dec 18, 2020, at 10:31, Koen Derks <notifications@github.com> wrote:
>
> ﻿
> This is my screen with the abovementioned options, yours is not the same I imagine?
>
>
>
