### 2020-12-18 #2733 @koenderks
Yes, I'm sorry I pasted the wrong screen unfortunately, this is the correct one. I think the variable "bookValues" must be placed under the header "Ist position" and the variable "auditValues" under the header "Soll position" to get the correct output.

![image](https://user-images.githubusercontent.com/25059399/102632393-53422e00-414f-11eb-8703-797b535a2f8b.png)
