### 2020-12-18 #2733 @trstewart212
Oh, okay thanks. “Soll” is not a term used in the US or the Anglo auditing world, but I do know it from my work in the Netherlands and should have known that it means auditValue in this context. Just forgot. My other questions remain.
Trevor

Trevor Stewart
175 West 13th Street, Apt 19D
New York, NY 10011
+1(212)929-0321
+1(917)402-9800 iPhone
trsny@verizon.net
