### 2020-12-23 #2733 @koenderks
Hi Trevor,

Thank you for mentioning this. I have chosen the names "Ist" and "Soll" but these things are easy to adjust. If you have any feedback or suggestions about the naming in the software I would be pleased to hear that!

> The note says that the upper bound is calculated according to the Stringer method with LTA adjustment. As a reviewer, how do I know this has been correctly implemented, mathematically?

Within JASP, the users' options are gathered and passed to the R package `jfa` (version JASP-v0.5.0) whose source code can be found at https://github.com/koenderks/jfa. This R package performs all the relevant calculations and then passes the results back to JASP. This makes sure that the methods and calculations are implemented in a single place, and they are documented in the [manual](https://github.com/koenderks/jfa/raw/master/man/manual/jfa_0.5.0.pdf) of the `jfa` package.

At the time of writing, there are no explicit formula's available in the manual that outline how the calculations are made specifically, but raw code can be found in the file [.interna.R](https://github.com/koenderks/jfa/blob/master/R/internal.R). Instead, in the manual I refer to the literature from which these methods are coming. It is of course possible to add these formulas to the manual.

> The evaluation refers to prior information. What is the prior, and where does it come from?

This is indeed a mistake. This text should only be shown in the Bayesian analysis, where a prior distribution is specified according to the referenced prior information. This text is shared between the Bayesian and frequentist analyses, so I will have to adjust that so that they provide unique texts based on the type of analysis.

Happy holidays!
