### 2021-02-17 #2733 @trstewart212
Yes, Koen did answer my questions but I don’t know whether the points I made are being addressed. In any case, my questions were based on just a cursory look as I was just trying to figure out at a very basic level how JFA works.

Since then I have made several attempts to get back into this but every time find myself stymied by the absence of a comprehensive cohesive paper that describes the auditing problems being addressed, the mathematical solutions, and their implementation in JFA. What I have seen includes:
A one page very high level Summary.
A 19-page manual that describes some of the functions and defines some probability distributions I would expect in Bayesian work.
I don’t see any write-up of how and why the mathematics works and there are no worked examples to give me an understanding. In the absence of these things I don’t see what I can add as a reviewer of the functionality provided by JFA.

If I had documentation as I have described and if Koen or someone were to work through the review checklist and describe the extent to which JFA satisfies the requirements, supported by reviewable evidence, then I might be able to make some headway.

Please understand that the above is not a commentary on the quality of the submission but on its reviewability.

Best,

Trevor

Trevor Stewart
