### 2021-03-11 beb4252
Add JOSS paper to repository
