+++ paper.bib
@@ -1,0 +1,71 @@
+@misc{TheWorldBank:2019, 
+    title={Listed Domestic Companies}, 
+    author={{The World Bank}},
+    year = {2019},
+    url={https://data.worldbank.org/indicator/cm.mkt.ldom.no}, 
+    note = {Accessed: 10-01-2020}
+}
+@MISC{JASP:2018,
+AUTHOR = {{JASP Team}},
+TITLE = {{JASP (Version 0.11.1)[Computer software]}},
+YEAR = {2019},
+URL = {https://jasp-stats.org/}
+}
+@article{george:2012,
+  title={How much does IFRS cost? IFRS adoption and audit fees},
+  author={De George, Emmanuel T and Ferguson, Colin B and Spear, Nasser A},
+  journal={The Accounting Review},
+  volume={88},
+  number={2},
+  pages={429--462},
+  year={2012},
+  publisher={American Accounting Association}
+}
+@article{appelbaum:2017,
+  title={Big Data and analytics in the modern audit engagement: Research needs},
+  author={Appelbaum, D and Kogan, A and Vasarhelyi, M A},
+  journal={Auditing: A Journal of Practice \& Theory},
+  volume={36},
+  number={4},
+  pages={1--27},
+  year={2017},
+  publisher={American Accounting Association}
+}
+@Manual{RTeam:2019,
+  title = {R: A language and environment for statistical computing},
+  author = {{R Core Team}},
+  organization = {R Foundation for Statistical Computing},
+  address = {Vienna, Austria},
+  year = {2019},
+  url = {https://www.R-project.org/},
+}
+@article{love:2019,
+  title={JASP: graphical statistical software for common statistical designs},
+  author={Love, J and Selker, R and Marsman, M and Jamil, T and Dropmann, D and Verhagen, J and Ly, A and Gronau, Q F and Smira, M and Epskamp, S and Matzke, D and Wild, A and Knight, P and Rouder J N, and Morey, R D, and Wagenmakers, E -J},
+  journal={Journal of Statistical Software},
+  volume={88},
+  number={2},
+  year={2019}
+}
+@misc{IFAC:2010,
+  title={International standard on auditing 530: Audit sampling},
+  author={{International Federation of Accountants}},
+  year={2010},
+  url={https://www.ifac.org/system/files/downloads/a027-2010-iaasb-handbook-isa-530.pdf}, 
+  note = {Accessed: 10-01-2020}
+}
+@Manual{derks:2020a,
+  title = {{jfa}: Bayesian and Classical Audit Sampling},
+  author = {Derks, K},
+  year = {2020},
+  note = {R package version 0.1.0},
+  url = {https://CRAN.R-project.org/package=jfa},
+}
+@article{derks:2020b,
+  title={Priors in a {B}ayesian audit: How integration of existing knowledge into the prior distribution can increase transparency, efficiency, and quality},
+  author={Derks, K and {van Batenburg}, P and {de Swart}, J and Wagenmakers, E -J and Wetzels, R},
+  journal={Submitted},
+  volume={},
+  number={},
+  year={2020}
+}
