+++ paper.bib
@@ -1,0 +1,77 @@
+@misc{TheWorldBank:2019, 
+    title={Listed Domestic Companies}, 
+    author={{The World Bank}},
+    year = {2019},
+    url={https://data.worldbank.org/indicator/cm.mkt.ldom.no}, 
+    note = {Accessed: 10-01-2020}
+}
+@MISC{JASP:2018,
+AUTHOR = {{JASP Team}},
+TITLE = {{JASP (Version 0.15.0)[Computer software]}},
+YEAR = {2021},
+URL = {https://jasp-stats.org/}
+}
+@article{george:2012,
+  title={How much does IFRS cost? IFRS adoption and audit fees},
+  author={De George, Emmanuel T and Ferguson, Colin B and Spear, Nasser A},
+  journal={The Accounting Review},
+  volume={88},
+  number={2},
+  pages={429--462},
+  year={2012},
+  publisher={American Accounting Association},
+  doi={10.2308/accr-50317}
+}
+@article{appelbaum:2017,
+  title={Big Data and analytics in the modern audit engagement: Research needs},
+  author={Appelbaum, D and Kogan, A and Vasarhelyi, M A},
+  journal={Auditing: A Journal of Practice \& Theory},
+  volume={36},
+  number={4},
+  pages={1--27},
+  year={2017},
+  publisher={American Accounting Association},
+  doi={10.2308/ajpt-51684}
+}
+@Manual{RTeam:2019,
+  title = {R: A language and environment for statistical computing},
+  author = {{R Core Team}},
+  organization = {R Foundation for Statistical Computing},
+  address = {Vienna, Austria},
+  year = {2021},
+  url = {https://www.R-project.org/},
+}
+@article{love:2019,
+  title={JASP: graphical statistical software for common statistical designs},
+  author={Love, J and Selker, R and Marsman, M and Jamil, T and Dropmann, D and Verhagen, J and Ly, A and Gronau, Q F and Smira, M and Epskamp, S and Matzke, D and Wild, A and Knight, P and Rouder J N, and Morey, R D, and Wagenmakers, E -J},
+  journal={Journal of Statistical Software},
+  volume={88},
+  number={2},
+  year={2019},
+  doi={10.18637/jss.v088.i02}
+}
+@book{IAASB:2018,
+  title={Handbook of international quality control, auditing review, other assurance,
+and related services pronouncements ({V}ol. {I})},
+  author={{International Auditing and Assurance Standards Board (IAASB)}},
+  year={2018},
+  publisher = {{International Federation of Accountants}},
+  address = {New York}
+}
+@Manual{derks:2020a,
+  title = {{jfa}: Bayesian and Classical Audit Sampling},
+  author = {Derks, K},
+  year = {2021},
+  note = {R package version 0.6.0},
+  url = {https://CRAN.R-project.org/package=jfa},
+}
+@article{derks:2020b,
+  title={Priors in a {B}ayesian audit: How integration of existing information into the prior distribution can improve audit transparency and efficiency},
+  author={Derks, K and {van Batenburg}, P and {de Swart}, J and Wagenmakers, E -J and Wetzels, R},
+  journal={International Journal of Auditing},
+  volume={25},
+  number={3},
+  pages = {621--636},
+  year={2021},
+  doi={10.1111/ijau.12240}
+}
