### 2021-03-12 #2733 @koenderks
Dear @drvinceknight, dear @trstewart212,

Recently I have split off all the files that are included in JASP for Audit (the current submission) from the remaining code in JASP. As there is now a completely separate repository specifically for JASP for Audit, I was wondering if it is possible to change the submission repository from https://github.com/koenderks/jasp-desktop (branch JfA-JOSS) to the new repository https://github.com/koenderks/jaspAudit (branch JOSS)? This might make it more clear which files are in the scope of the current submission. Would this be possible @drvinceknight?

In this new repository, I have also added a user manual for the software, which @trstewart212 and @fgeertman may find helpful in reviewing the software. The user manual itself can also be via this [link](https://github.com/jasp-stats/jaspAudit/raw/master/man/manual.pdf). It contains a layperson's introduction to statistical auditing and the sampling workflow. It then proceeds to discuss how this sampling workflow is implemented in JASP for Audit using illustrated examples.

Moreover, in both the manual and the repository for JASP for Audit itself, I have made a clear reference to the calculation and validation of the statistical results via the `jfa` R package. This R package performs the calculations underlying JASP for Audit. Its source code, as well as more information about the validation of the statistical results, can be found in the repository of the package at https://github.com/koenderks/jfa. In this repository, I have benchmarked `jfa`'s statistical results to well-known sources such as [AICPA (2017) - Appendix A](https://onlinelibrary.wiley.com/doi/pdf/10.1002/9781119448617.app1) and [AICPA (2017) - Appendix C](https://onlinelibrary.wiley.com/doi/pdf/10.1002/9781119448617.app3). @trstewart212 I hope that these resources are able to give you a better indication of the verification of the mathematical solutions.

@trstewart212 I hope that this can help you get started in JASP for Audit. If you encounter any questions that I can clear up (maybe even by showing them in JASP for Audit to you), please let me know!

In the new repository is the current code for JASP for Audit, in which I have incorporated the feedback by @trstewart212. Please note that this code is not "live" yet since it is not released on the official website for JASP (www.jasp-stats.org). However, a development version of the program containing these changes can be downloaded [here for Windows](https://static.jasp-stats.org/Nightlies/Windows/JASP-nightly-development-7ffa0edb75794607dc11635e292fe41a50be3513.msi) and [here for MAC](https://static.jasp-stats.org/Nightlies/MacOS/JASP-nightly-development-7ffa0edb75794607dc11635e292fe41a50be3513.dmg).

Sincerely,
Koen
