### 2021-03-24 #2733 @drvinceknight
@arfon :

Can the repo be moved:

**from** https://github.com/koenderks/jasp-desktop (branch JfA-JOSS)
**to** https://github.com/koenderks/jaspAudit (branch JOSS)
