### 2021-06-18 #2733 @fgeertman
Review Question:  •	Installation: Does installation proceed as outlined in the documentation?

JASP for Audit is part of JASP. JASP is to download from https://jasp-stats.org/ . No issues noted at installation. It is however recommended to refer to JASP, and where it can be downloaded and how the Audit module can be made (in)visible
