### 2021-06-18 #2733 @drvinceknight
> I noticed that the link under `repository url` in the review checklist still directs to https://github.com/koenderks/jasp-desktop. Is it possible to adjust this to the current repository url https://github.com/koenderks/jaspAudit/tree/JOSS?

I believe I have now modified this correctly.
