+++ paper.bib
@@ -7,8 +7,8 @@
 }
 @MISC{JASP:2018,
 AUTHOR = {{JASP Team}},
-TITLE = {{JASP (Version 0.11.1)[Computer software]}},
-YEAR = {2019},
+TITLE = {{JASP (Version 0.14.1)[Computer software]}},
+YEAR = {2021},
 URL = {https://jasp-stats.org/}
 }
 @article{george:2012,
@@ -36,7 +36,7 @@ URL = {https://jasp-stats.org/}
   author = {{R Core Team}},
   organization = {R Foundation for Statistical Computing},
   address = {Vienna, Austria},
-  year = {2019},
+  year = {2021},
   url = {https://www.R-project.org/},
 }
 @article{love:2019,
@@ -47,25 +47,27 @@ URL = {https://jasp-stats.org/}
   number={2},
   year={2019}
 }
-@misc{IFAC:2010,
-  title={International standard on auditing 530: Audit sampling},
-  author={{International Federation of Accountants}},
-  year={2010},
-  url={https://www.ifac.org/system/files/downloads/a027-2010-iaasb-handbook-isa-530.pdf}, 
-  note = {Accessed: 10-01-2020}
+@book{IAASB:2018,
+  title={Handbook of international quality control, auditing review, other assurance,
+and related services pronouncements ({V}ol. {I})},
+  author={{International Auditing and Assurance Standards Board (IAASB)}},
+  year={2018},
+  publisher = {{International Federation of Accountants}},
+  address = {New York}
 }
 @Manual{derks:2020a,
   title = {{jfa}: Bayesian and Classical Audit Sampling},
   author = {Derks, K},
-  year = {2020},
-  note = {R package version 0.1.0},
+  year = {2021},
+  note = {R package version 0.5.5},
   url = {https://CRAN.R-project.org/package=jfa},
 }
 @article{derks:2020b,
-  title={Priors in a {B}ayesian audit: How integration of existing knowledge into the prior distribution can increase transparency, efficiency, and quality},
+  title={Priors in a {B}ayesian audit: How integration of existing information into the prior distribution can improve audit transparency and efficiency},
   author={Derks, K and {van Batenburg}, P and {de Swart}, J and Wagenmakers, E -J and Wetzels, R},
-  journal={Submitted},
+  journal={International Journal of Auditing},
   volume={},
   number={},
-  year={2020}
+  pages = {1--16},
+  year={2021}
 }
