+++ paper.md
@@ -31,8 +31,8 @@ bibliography: paper.bib
 
 In many countries, listed organizations are required to be audited by law. In the United States alone there were 4,397 listed companies that had at least one audit performed in 2017 [@TheWorldBank:2019]. Most audits are complex and come with high quality requirements, and the associated costs can be high [@george:2012]. Because the outcome of an audit is a probabilistic statement (for example, with 95% certainty the general ledger contains no material misstatements), much research has been done in how statistical techniques can increase audit quality and reduce audit complexity [@appelbaum:2017]. JASP for Audit (``JfA``) was developed to facilitate research into Bayesian statistics in an audit context and to simplify the statistical aspects of an audit.
 
-``JfA`` is a module for the free and open-source statistical software platform JASP [@love:2019; @JASP:2018]. The module consists of the ``R`` [@RTeam:2019] package ``jfa`` [@derks:2020a] and a graphical user interface around the ``jfa`` package that provides a continuous workflow for the auditor, and automatically creates an audit report containing the results, visualizations, and statistical interpretation of these results. The graphical user interface was implemented in ``QML`` to create an interactive layout that dynamically responds to user input. This simple point-and-click layout was designed with the auditor in mind, which means that relevant options are highlighted clearly and advanced options are hidden by default. The user input is sent to the ``jfa`` package in ``R`` for fast and efficient computing of the results. The results are returned to the ``JfA`` module and displayed in such a way (i.e., with explanatory text that explains the results in non-statistical terms) that any auditor, student, and researcher, can understand the statistical theory underlying their results. In addition, the integration in ``R`` enables ``JfA`` to easily import high-quality visualization packages to create reportable tables and figures that clarify the statistical results. This approach minimizes the dependency on an auditor’s statistical knowledge so that any auditor can use ``JfA``.
+``JfA`` is a module for the free and open-source statistical software platform JASP [@love:2019; @JASP:2018]. The module consists of the ``R`` [@RTeam:2019] package ``jfa`` [@derks:2020a] and a graphical user interface around the ``jfa`` package that provides a continuous workflow for the auditor, and supports audit documentation by automatically creating a report containing the results, visualizations, and statistical interpretation of these results. The graphical user interface was implemented in ``QML`` to create an interactive layout that dynamically responds to user input. This simple point-and-click layout was designed with the auditor in mind, which means that relevant options are highlighted clearly and advanced options are hidden by default. The user input is sent to the ``jfa`` package in ``R`` for fast and efficient computing of the results. The results are returned to the ``JfA`` module and displayed in such a way (i.e., with explanatory text that explains the results in non-statistical terms) that any auditor, student, and researcher, can understand the statistical theory underlying their results. In addition, the integration in ``R`` enables ``JfA`` to easily import high-quality visualization packages to create reportable tables and figures that clarify the statistical results. This approach minimizes the dependency on an auditor’s statistical knowledge so that any auditor can use ``JfA``.
 
[…]
