### 2021-06-19 #2733 @koenderks
Hi @drvinceknight,

It seems like whedon currently cannot find any papers to compile in the submission repository. Do you have any tips on how to resolve this issue? Many thanks in advance.

@fgeertman I will extend the manual so that it contains information on how/where to download JASP and how to make the audit module (in)visible.
