### 2021-06-23 #2733 @koenderks
Thank you for the fix @drvinceknight and @danielskatz

@fgeertman I have extended the [software manual](https://github.com/jasp-stats/jaspAudit/raw/master/man/manual.pdf) to include an explanation to downloading JASP and making the Audit module (in)visible.
