### 2021-07-09 #2733 @fgeertman
Hi,

I did and asked some questions to @koenderks<https://github.com/koenderks>, which I will follow-up after summer break.

Regards,

Ferry

From: Vince Knight ***@***.***
Sent: Friday, July 9, 2021 12:00 PM
To: openjournals/joss-reviews
Cc: Ferry Geertman; Mention
Subject: Re: [openjournals/joss-reviews] [REVIEW]: JASP for Audit: Bayesian Tools for the Auditing Practice (#2733)

@fgeertman<https://github.com/fgeertman> when you have time, it would be great, if you could take a look at the modification made by @koenderks<https://github.com/koenderks> and let us know how the review progresses as a result of that. Thanks.
