### 2021-09-30 #2733 @fgeertman
Yes, will do next week.

Thanks

From: Vince Knight ***@***.***
Sent: Monday, September 27, 2021 3:48 PM
To: openjournals/joss-reviews
Cc: Ferry Geertman; Mention
Subject: Re: [openjournals/joss-reviews] [REVIEW]: JASP for Audit: Bayesian Tools for the Auditing Practice (#2733)

@fgeertman<https://github.com/fgeertman> just a polite nudge in case you have time now to look at this.
