### 2021-10-04 #2733 @trstewart212
Have the outstanding items identified in the attached July 8 email been addressed?

Trevor Stewart
175 West 13th Street, Apt 19D
New York, NY 10011
+1(212)929-0321
+1(917)402-9800 iPhone
***@***.***
