### 2021-10-05 #2733 @koenderks
Hi all,

In the attached document ([Comments.docx](https://github.com/openjournals/joss-reviews/files/7283797/Review.JASP.-.KD.docx)) is my response to the July 8 email. Here I have tried to address the issues raised as best as possible, and have tried to explain how and where the requested information can be located online. Can this information be of any help in the review process? I would of course be open to a discussion on how to tackle the remaining challenges in the review.

Note that JASP 0.15 is now available from [www.jasp-stats.org](www.jasp-stats.org), which contains an updated version of JASP for audit / the audit module.

Sincerely,
Koen
