+++ paper.bib
@@ -7,7 +7,7 @@
 }
 @MISC{JASP:2018,
 AUTHOR = {{JASP Team}},
-TITLE = {{JASP (Version 0.14.1)[Computer software]}},
+TITLE = {{JASP (Version 0.15.0)[Computer software]}},
 YEAR = {2021},
 URL = {https://jasp-stats.org/}
 }
@@ -19,7 +19,8 @@ URL = {https://jasp-stats.org/}
   number={2},
   pages={429--462},
   year={2012},
-  publisher={American Accounting Association}
+  publisher={American Accounting Association},
+  doi={10.2308/accr-50317}
 }
 @article{appelbaum:2017,
   title={Big Data and analytics in the modern audit engagement: Research needs},
@@ -29,7 +30,8 @@ URL = {https://jasp-stats.org/}
   number={4},
   pages={1--27},
   year={2017},
-  publisher={American Accounting Association}
+  publisher={American Accounting Association},
+  doi={10.2308/ajpt-51684}
 }
 @Manual{RTeam:2019,
   title = {R: A language and environment for statistical computing},
@@ -45,7 +47,8 @@ URL = {https://jasp-stats.org/}
   journal={Journal of Statistical Software},
   volume={88},
   number={2},
-  year={2019}
+  year={2019},
+  doi={10.18637/jss.v088.i02}
 }
 @book{IAASB:2018,
   title={Handbook of international quality control, auditing review, other assurance,
@@ -59,15 +62,16 @@ and related services pronouncements ({V}ol. {I})},
   title = {{jfa}: Bayesian and Classical Audit Sampling},
   author = {Derks, K},
   year = {2021},
-  note = {R package version 0.5.5},
+  note = {R package version 0.6.0},
   url = {https://CRAN.R-project.org/package=jfa},
 }
 @article{derks:2020b,
   title={Priors in a {B}ayesian audit: How integration of existing information into the prior distribution can improve audit transparency and efficiency},
   author={Derks, K and {van Batenburg}, P and {de Swart}, J and Wagenmakers, E -J and Wetzels, R},
   journal={International Journal of Auditing},
-  volume={},
-  number={},
-  pages = {1--16},
-  year={2021}
+  volume={25},
+  number={3},
+  pages = {621--636},
+  year={2021},
+  doi={10.1111/ijau.12240}
 }
