### 2021-11-12 #2733 @fgeertman
Hi, There is no need for you to apologize given the time it took us to review ☺

From: Vince Knight ***@***.***
Sent: Friday, November 12, 2021 12:45 PM
To: openjournals/joss-reviews
Cc: Ferry Geertman; Mention
Subject: Re: [openjournals/joss-reviews] [REVIEW]: JASP for Audit: Bayesian Tools for the Auditing Practice (#2733)

Apologies for the delay. Everything looks good to me.
