### 2021-11-13 #2733 @koenderks
Hi @drvinceknight,

- I have tagged the submitted version `jaspAudit-v0.16` on GitHub, see https://github.com/koenderks/jaspAudit/releases/tag/jaspAudit-v0.16.
- I have archived the software on Zenodo at [https://doi.org/10.5281/zenodo.5690631](https://doi.org/10.5281/zenodo.5690631) with the DOI of `10.5281/zenodo.5690631` and version number `0.16`.
