### 2021-12-02 #2733 @koenderks
Thank you @kthyng, @drvinceknight, @trstewart212, and @fgeertman for all the work that has gone into this review and publication process. Your time and effort is very much appreciated!

@drvinceknight I have one remaining question regarding the [`JOSS`](https://github.com/koenderks/jaspAudit/tree/JOSS) branch. Is it safe to remove this branch or is there a need to keep it around?
