### 2021-12-04 #2733 @drvinceknight
Hi @koenderks, I'm not actually sure here. I suggest you remove it and if that causes anything to fall over we can bring it back.
