### 2021-12-05 #2733 @arfon
> @drvinceknight I have one remaining question regarding the JOSS branch. Is it safe to remove this branch or is there a need to keep it around?

Nope, you can delete it.
